third_party/blink/web_tests/virtual/scoped-custom-element-registry/external/wpt/custom-elements/registries/Element-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] customElementRegistry on an element inside a declarative shadow DOM with shadowrootcustomelementregistry should return the scoped registry after calling initialize
  assert_equals: expected object "[object CustomElementRegistry]" but got null
Harness: the test ran to completion.
